hikvision-doorbell/requirements.txt
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
flake8==6.0.0
goodconf==3.0.0
ha-mqtt-discoverable==0.6.0
idna==3.4
iniconfig==2.0.0
loguru==0.6.0
mccabe==0.7.0
packaging==23.0
paho-mqtt==1.6.1
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
pytest==7.2.1
pytest-mock==3.10.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
